Adds line drawing and IO rendering
Adds scripts responsible for instantiating the input/output nodes as well as drawing the lines from one node to another.

Bugs:
Can't end a line on a node, just starts a new line
I want to be able to draw a line with multiple bends in it. Currently a line is only straight and can't be bendy.