Edit gitignore to whitelist scripts
Changes the .gitignore file to only whitelist the scripts.
It's gotten to the point where managing the repo puts a lot of strain on github desktop and by extension slows down my computer. By changing the structure of the git ignore to only whitelist my scripts, I will lessen the burden on github desktop an on my computer's CPU.

I also had to change the folder name to have underscores instead of spaces and this makes it look like all my scripts are being removed and added back.