Add rewiring with rules
With this commit you can choose to rewire the input or output of each connecting wire with the constraint that
1. Two inputs can't be directly connected with eachother
2. Two outputs can't be directly connected with eachother

You also can not rewire two wires at once.

In the process of making this commit, I broke the logic propagation commit but what is essential to remember going forward is that the output of a chip is the input to wire and the output to a wire is the input to a chip.